further work on package dialogs
