+++ catalog/app/containers/Bucket/PackageDialog/index.ts
@@ -3,6 +3,7 @@ export { default as DialogLoading } from './DialogLoading'
 export { default as DialogSuccess } from './DialogSuccess'
 export * from './FilesInput'
 export * from './Layout'
+export * from './PackageCreationForm'
 export * from './PackageDialog'
 export * from './Skeleton'
 export { default as SubmitSpinner } from './SubmitSpinner'
